Update README.md
The Hotel Booking Application is a meticulously crafted platform designed to streamline the process of finding and booking hotel rooms. This project aims to enhance the user experience by providing a visually appealing, user-friendly interface that caters to the needs of modern travelers. The application offers a comprehensive set of features, including a responsive design that adapts seamlessly to various devices, ensuring accessibility and convenience for all users. By integrating essential functionalities such as room browsing, booking management, and customer support, the application strives to deliver a hassle-free and enjoyable booking experience.